@@ -9,53 +9,38 @@
         "url": "https://lucapattocchio.dev"
     },
     "license": "MIT",
+    "type": "module",
     "scripts": {
-        "start": "vite",
-        "build": "tsc && vite build",
-        "preview": "vite preview",
-        "lint": "eslint -c .eslintrc.json --ext .js,.jsx .",
-        "lint:fix": "npm run lint -- --fix"
+        "dev": "vite",
+        "build": "tsc -b && vite build",
+        "lint": "eslint .",
+        "preview": "vite preview"
     },
     "dependencies": {
-        "@vitejs/plugin-react": "4.3.3",
-        "react": "18.3.1",
-        "react-dom": "18.3.1",
-        "typescript": "5.6.3",
-        "vite": "5.4.10",
-        "vite-tsconfig-paths": "5.1.2"
+        "@heroicons/react": "^2.2.0",
+        "clsx": "^2.1.1",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-router": "^7.0.1",
+        "react-use": "^17.5.1"
     },
     "devDependencies": {
-        "@types/node": "20.14.9",
-        "@types/react": "18.3.12",
-        "@types/react-dom": "18.3.1",
-        "@typescript-eslint/eslint-plugin": "8.13.0",
-        "@typescript-eslint/parser": "8.13.0",
-        "autoprefixer": "10.4.20",
-        "eslint": "8.57.1",
-        "eslint-config-prettier": "9.1.0",
-        "eslint-plugin-react": "7.37.2",
-        "eslint-plugin-react-hooks": "5.0.0",
-        "postcss": "8.4.48",
-        "prettier": "3.3.3",
-        "prettier-plugin-tailwindcss": "0.6.8",
-        "tailwindcss": "3.4.14"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+        "@eslint/js": "^9.15.0",
+        "@types/react": "^18.3.12",
+        "@types/react-dom": "^18.3.1",
+        "@vitejs/plugin-react": "^4.3.4",
+        "autoprefixer": "^10.4.20",
+        "eslint": "^9.15.0",
+        "eslint-plugin-prettier": "^5.2.1",
+        "eslint-plugin-react-hooks": "^5.0.0",
+        "eslint-plugin-react-refresh": "^0.4.14",
+        "globals": "^15.12.0",
+        "postcss": "^8.4.49",
+        "prettier": "^3.4.1",
+        "prettier-plugin-tailwindcss": "^0.6.9",
+        "tailwindcss": "^3.4.15",
+        "typescript": "~5.6.2",
+        "typescript-eslint": "^8.15.0",
+        "vite": "^6.0.1"
     }
 }
